pyproject.toml
[tool.poetry]
name = "cycle"
version = "0.1.0"
description = "MAC0467 - MVP"
authors = ["ArthurPilone <art.pilone@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
twilio = "^8.1.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
